backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
google-genai
supabase
python-dotenv
requests
pydantic
httpx
